AP-1350 Read pagenumber and pages results per page from request. Do not duplicate filterqueries
git-svn-id: https://opensource.kantega.no/svn/openaksess/branches/openaksess-solr@3900 fd808399-8219-4f14-9d4c-37719d9ec93d

+++ modules/search-api/src/java/no/kantega/search/api/search/SearchResponse.java
@@ -81,7 +81,7 @@ public class SearchResponse {
     }
 
     public int getNumberOfPages() {
-        return (int) Math.ceil(numberOfHits.intValue() / query.getResultsPerPage());
+        return (int) Math.ceil(numberOfHits.doubleValue() / query.getResultsPerPage());
     }
 
     public void setFacets(Map<String,Collection<FacetResult>> facets) {
